backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
uwsgi==2.0.22